requirements.txt
streamlit 		# obviously
chromadb 		# try with this first. alternative: faiss?
openai 			# for APi access to openrouter
sentence-transformers

python-dotenv		# let's see if this the bearer error? 
